+++ src/m365/aad/commands/o365group/o365group-recyclebinitem-restore.ts
@@ -16,19 +16,17 @@ interface Options extends GlobalOptions {
   id: string;
 }
 
-class AadO365GroupRestoreCommand extends GraphCommand {
+class AadO365GroupRecycleBinItemRestoreCommand extends GraphCommand {
   public get name(): string {
-    return commands.O365GROUP_RESTORE;
+    return commands.O365GROUP_RECYCLEBINITEM_RESTORE;
   }
 
   public get description(): string {
     return 'Restores a deleted Microsoft 365 Group';
   }
 
-  public getTelemetryProperties(args: CommandArgs): any {
-    const telemetryProps: any = super.getTelemetryProperties(args);
-    telemetryProps.id = (!(!args.options.id)).toString();
-    return telemetryProps;
+  public alias(): string[] | undefined {
+    return [commands.O365GROUP_RESTORE];
   }
 
   public commandAction(logger: Logger, args: CommandArgs, cb: () => void): void {
@@ -39,8 +37,10 @@ class AadO365GroupRestoreCommand extends GraphCommand {
     const requestOptions: any = {
       url: `${this.resource}/v1.0/directory/deleteditems/${args.options.id}/restore/`,
       headers: {
-        'accept': 'application/json;odata.metadata=none'
-      }
+        'accept': 'application/json;odata.metadata=none',
+        'content-type': 'application/json'
+      },
+      responseType: 'json'
     };
 
     request
@@ -68,4 +68,4 @@ class AadO365GroupRestoreCommand extends GraphCommand {
   }
 }
 
-module.exports = new AadO365GroupRestoreCommand();
+module.exports = new AadO365GroupRecycleBinItemRestoreCommand();