+++ src/m365/aad/commands/siteclassification/siteclassification-disable.ts
@@ -35,7 +35,7 @@ class AadSiteClassificationDisableCommand extends GraphCommand {
   public commandAction(logger: Logger, args: CommandArgs, cb: (err?: any) => void): void {
     const disableSiteClassification: () => void = (): void => {
       const requestOptions: any = {
-        url: `${this.resource}/beta/settings`,
+        url: `${this.resource}/v1.0/groupSettings`,
         headers: {
           accept: 'application/json;odata.metadata=none'
         },
@@ -63,7 +63,7 @@ class AadSiteClassificationDisableCommand extends GraphCommand {
           }
 
           const requestOptions: any = {
-            url: `${this.resource}/beta/settings/` + unifiedGroupSetting[0].id,
+            url: `${this.resource}/v1.0/groupSettings/` + unifiedGroupSetting[0].id,
             headers: {
               accept: 'application/json;odata.metadata=none',
               'content-type': 'application/json'