+++ src/m365/spo/commands/listitem/listitem-get.ts
@@ -19,7 +19,7 @@ interface Options extends GlobalOptions {
   listId?: string;
   listTitle?: string;
   id: string;
-  fields?: string;
+  properties?: string;
 }
 
 class SpoListItemGetCommand extends SpoCommand {
@@ -49,8 +49,8 @@ class SpoListItemGetCommand extends SpoCommand {
       `${args.options.webUrl}/_api/web/lists(guid'${encodeURIComponent(listIdArgument)}')`
       : `${args.options.webUrl}/_api/web/lists/getByTitle('${encodeURIComponent(listTitleArgument)}')`);
 
-    const fieldSelect: string = args.options.fields ?
-      `?$select=${encodeURIComponent(args.options.fields)}` :
+    const propertiesSelect: string = args.options.properties ?
+      `?$select=${encodeURIComponent(args.options.properties)}` :
       (
         (!args.options.output || args.options.output === 'text') ?
           `?$select=Id,Title` :
@@ -58,7 +58,7 @@ class SpoListItemGetCommand extends SpoCommand {
       );
 
     const requestOptions: any = {
-      url: `${listRestUrl}/items(${args.options.id})${fieldSelect}`,
+      url: `${listRestUrl}/items(${args.options.id})${propertiesSelect}`,
       headers: {
         'accept': 'application/json;odata=nometadata'
       },
@@ -89,7 +89,7 @@ class SpoListItemGetCommand extends SpoCommand {
         option: '-t, --listTitle [listTitle]'
       },
       {
-        option: '-f, --fields [fields]'
+        option: '-p, --properties [properties]'
       }
     ];
 
@@ -104,7 +104,7 @@ class SpoListItemGetCommand extends SpoCommand {
         'listId',
         'listTitle',
         'id',
-        'fields'
+        'properties'
       ]
     };
   }